feat(many): see body for all changes
1. move kernel functions in Rasterizer pipeline into standalone file,
so that it is easy to modify kernel related codes or pipeline
related code.
2. rewrite rasterize kernel in Rasterizer pipeline. In the former way,
we spawn a thread for each fragment, but it renders so slow. Now this
kernel will only have one thread testing each triangles. For each
triangle, we separate n thread, n is the number of fragments in bounding
box of the triangle, and testing if the fragment is in the triangle and
update fragment data if neccessary.
3. add support for keyboard input: wsad, sapce and ctrl.